Improved: Moved common spec files into common/tests folder.
